refactor: move common library controls to the topbar
