+++ enterprise/frontend/src/metabase-enterprise/transforms-python/components/PythonTransformEditor/PythonTransformTopBar/PythonTransformTopBar.tsx
@@ -1,14 +1,25 @@
+import { t } from "ttag";
+
 import { hasFeature } from "metabase/admin/databases/utils";
 import {
   skipToken,
   useGetDatabaseQuery,
   useListDatabasesQuery,
 } from "metabase/api";
+import { ForwardRefLink } from "metabase/common/components/Link";
+import * as Urls from "metabase/lib/urls";
 import { DatabaseDataSelector } from "metabase/query_builder/components/DataSelector";
-import { Flex } from "metabase/ui";
+import { ActionIcon, Flex, Group, Icon, Tooltip } from "metabase/ui";
 import { EditDefinitionButton } from "metabase-enterprise/transforms/components/TransformEditor/EditDefinitionButton";
 import { doesDatabaseSupportTransforms } from "metabase-enterprise/transforms/utils";
+import { SHARED_LIB_IMPORT_PATH } from "metabase-enterprise/transforms-python/constants";
 import type { Database, DatabaseId, TransformId } from "metabase-types/api";
+
+import {
+  hasImport,
+  insertImport,
+  removeImport,
+} from "../PythonEditorBody/utils";
 
 import S from "./PythonTransformTopBar.module.css";
 
@@ -16,15 +27,21 @@ type PythonTransformTopBarProps = {
   databaseId?: DatabaseId;
   readOnly?: boolean;
   transformId?: TransformId;
+  source?: string;
   onDatabaseChange?: (databaseId: DatabaseId) => void;
+  onSourceChange?: (source: string) => void;
 };
 
 export function PythonTransformTopBar({
   databaseId,
   readOnly,
   transformId,
+  source,
   onDatabaseChange,
+  onSourceChange,
 }: PythonTransformTopBarProps) {
+  const showReadonlyControls = readOnly && transformId;
+
   const { data: database } = useGetDatabaseQuery(
     databaseId != null ? { id: databaseId } : skipToken,
   );
@@ -34,6 +51,16 @@ export function PythonTransformTopBar({
     const newDatabaseId = value ? parseInt(value) : undefined;
     if (newDatabaseId != null && newDatabaseId !== databaseId) {
       onDatabaseChange?.(newDatabaseId);
+    }
+  };
+
+  const handleToggleSharedLib = () => {
+    if (source != null && onSourceChange) {
+      if (hasImport(source, SHARED_LIB_IMPORT_PATH)) {
+        onSourceChange(removeImport(source, SHARED_LIB_IMPORT_PATH));
+      } else {
+        onSourceChange(insertImport(source, SHARED_LIB_IMPORT_PATH));
+      }
     }
   };
 
@@ -69,8 +96,8 @@ export function PythonTransformTopBar({
           />
         </Flex>
       )}
-      {readOnly && transformId && (
-        <Flex ml="auto" mr="lg" align="center" h="3rem">
+      <Flex ml="auto" mr="lg" align="center" h="3rem">
+        {showReadonlyControls ? (
           <EditDefinitionButton
             bg="transparent"
             fz="sm"
@@ -79,8 +106,31 @@ export function PythonTransformTopBar({
             size="xs"
             transformId={transformId}
           />
-        </Flex>
-      )}
+        ) : (
+          <Group gap="sm">
+            <Tooltip label={t`Import common library`}>
+              <ActionIcon
+                aria-label={t`Import common library`}
+                onClick={handleToggleSharedLib}
+              >
+                <Icon name="reference" c="text-primary" />
+              </ActionIcon>
+            </Tooltip>
+            <Tooltip label={t`Edit common library`}>
+              <ActionIcon
+                component={ForwardRefLink}
+                target="_blank"
+                aria-label={t`Edit common library`}
+                to={Urls.transformPythonLibrary({
+                  path: SHARED_LIB_IMPORT_PATH,
+                })}
+              >
+                <Icon name="pencil" c="text-primary" />
+              </ActionIcon>
+            </Tooltip>
+          </Group>
+        )}
+      </Flex>
     </Flex>
   );
 }
